+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change Log
 
+
 ## Added
 - open non-editable formats on views
 - formats for conversion docm, dot, dotx, epub, htm, html, otp, ots, ott, pot, potm, potx, pps, ppsm, ppsx,
  pptm, rtf, xlsm, xlt, xltm, xltx
+- keep intermediate versions when editing (forcesave)
+- version history with highlighting changes
 - detecting mobile browser
 
 ## 3.0.0